+++ cmd/cards/deck.go
@@ -1,0 +1,12 @@
+package main
+
+import "github.com/spf13/cobra"
+
+var deckCmd = &cobra.Command{
+	Use:   "deck",
+	Short: "Manage flashcard decks",
+}
+
+func init() {
+	rootCmd.AddCommand(deckCmd)
+}